No Errors. Fixed PWM. See diff. there is too much to write here. malloc was REMOVED since dynamic allocation is not supported in a 8-bit processor

+++ include/communication.h
@@ -20,7 +20,7 @@ void comm_flush(void); // Flush the buffers
 unsigned comm_go(void); // Begin transmission of the buffer
 unsigned comm_tx(unsigned char,unsigned); // Transmit a byte (add to TX FIFO) returns boolean
 unsigned char comm_rx(void); // Get a byte from the RX FIFO
-unsigned short comm_rx_word(void); // get a word from the RX FIFO
+unsigned int comm_rx_word(void); // get a word from the RX FIFO
 
 // Variables
 char comm_overflow;
